requirements.txt
# === AKIRA V21 ULTIMATE - Requirements ===
# Versão: Janeiro 2025
# Arquitetura: Multi-API com fallback + BART Emotion Analysis + Aprendizado Contínuo

# ============================================================
# 🔥 CORE - Web Framework & API
# ============================================================
flask>=2.0.0,<4.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
loguru>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ============================================================
# 🤖 IA & ML - Multi-API Support
# ============================================================
# Google Gemini API (google.genai - nova API)
google-genai>=1.0.0

# Mistral API via requests (sem cliente deprecated)
# A API é chamada diretamente via HTTP, sem precisar da biblioteca mistralai

# Groq API
groq>=0.4.0,<1.0.0

# Cohere API
cohere>=5.0.0,<6.0.0

# Together AI (compatible with OpenAI SDK)
openai>=1.0.0,<2.0.0

# HuggingFace Hub - versão COMPATÍVEL (não usa snapshot_download)
# ⚠️ IMPORTANTE: sentence-transformers 2.2.2+ não usa mais cached_download
huggingface-hub>=0.28.1

# Transformers core (BERT, BART, etc.) - MANTIDO A PEDIDO DO UTILIZADOR
transformers>=4.38.0,<4.50.0



# PyTorch - versão estável e compatível - MANTIDO A PEDIDO
torch>=2.1.0,<2.6.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Sentence Transformers - versão mais recente compatível - MANTIDO A PEDIDO
sentence-transformers>=2.2.2,<3.0.0

# ============================================================
# 📊 NLP & Text Processing
# ============================================================
numpy>=1.26.0,<2.0.0
nltk>=3.8.1,<3.10.0
spacy>=3.7.0,<4.0.0
# Modelos pt_core_news_lg serão baixados automaticamente
accelerate>=0.20.0

# ============================================================
# 🔌 Utilities & Integration
# ============================================================
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ============================================================
# 🗄️ Database
# ============================================================
# SQLite3 já vem com Python
# Opcional: Redis para cache (descomente se precisar)
# redis>=5.0.0

# ============================================================
# 📱 Web Search & Tools
# ============================================================
googlesearch-python>=1.1.0
ddgs>=5.3.0

# deepgram-sdk>=3.0.0  # Para STT (opcional)
# google-cloud-texttospeech>=2.0.0  # Para TTS (opcional)

# ============================================================
# 👁️ VISÃO COMPUTACIONAL & OCR
# ============================================================
# OpenCV headless - versão sem GUI, otimizada para servidores
opencv-python-headless>=4.8.0,<4.10.0

# Pillow - Manipulação de imagens PIL
pillow>=10.0.0,<11.0.0

# Tesseract OCR wrapper Python
pytesseract>=0.3.10

# Tesseract binary (Linux)
# Instalado via apt no Dockerfile:
# apt-get install tesseract-ocr tesseract-ocr-por tesseract-ocr-eng


# ============================================================
# 🎨 Media & QR Codes
# ============================================================
pillow>=10.0.0,<11.0.0
qrcode>=7.4.2,<8.0.0

# ============================================================
# ⚡ Performance & Monitoring
# ============================================================
# APScheduler>=7.0.0  # Para tarefas agendadas (opcional)
# prometheus-client>=0.19.0  # Para métricas (opcional)

# ============================================================
# 🧪 Testing (dev only - descomente se necessário)
# ============================================================
# pytest>=7.4.0
# pytest-cov>=4.1.0


# ============================================================
# 📝 NOTAS DE COMPATIBILIDADE (CRÍTICO!)
# ============================================================
#
# Para AKIRA V21 ULTIMATE:
# - sentence-transformers>=2.2.2 usa huggingface_hub.file_download
# - huggingface-hub>=0.23.0 não tem cached_download
# - Se tiver erro de import, atualize os pacotes:
#
# Exemplo de instalação limpa:
#   pip install --upgrade pip
#   pip install "sentence-transformers>=2.2.2"
#   pip install "huggingface-hub>=0.23.0,<0.27.0"
#   pip install "transformers>=4.38.0"
#   pip install "torch>=2.1.0"
#
# ============================================================

